IconIntegratedTitle moved out of helpers
