+++ src/helpers/words.js
@@ -1,5 +1,3 @@
-import React from 'react'
-
 /**
  * Splitting on spaces, convert a string to an array and remove the last element.
  * Will strip the last word from a string including punctuation.
@@ -17,20 +15,4 @@ export const getLastWord = (text) => {
     .trim()
     .split(' ')
     .pop()
-}
-
-/**
- * Use this to wrap things like icons in a span with the last word of a text string.
- * Doing so will prevent the icon from being isolated on a newline without any word.
- */
-export const IconIntegratedTitle = ({ title, children }) => {
-  return (
-    <>
-      <span>{removeLastWord(title)}</span>
-      <div style={{ display: 'inline-block' }}>
-        <span>{getLastWord(title)}</span>
-        {children}
-      </div>
-    </>
-  )
 }
